Enable development mode and add project watch directives
In the application.yaml configuration file, a flag for development mode was added to enable or disable development-specific settings. In addition, under deployment settings, watch directives were included for both 'classes' and 'resources'. This allows for easier tracking of code changes in these directories during development and should speed up the development process.