Add README file for the Portfolio Website
A README file has been added for the portfolio website. This file includes the website title, URL, and some basic instructions for building and running the server. The purpose of this change is to make the project more user-friendly and to provide clear information on how to interact with the project's build system and server.