Update Gradle dependencies for HTML templating
Added 'io.ktor.plugin' in plugins and two more dependencies for 'ktor-server-html-builder-jvm' and 'kotlin-css' in the templating's build.gradle.kts file. The new plugin and dependencies support the newly created BaseTemplate.kt to bolster the handling of HTML templates and standardize CSS across the application. This change was necessary to ensure consistency and improve readability and maintainability of the code.