Add BaseTemplate class for HTML templating
Introduced a new Kotlin file, BaseTemplate.kt, in the templating package. This file defines a BaseTemplate class that will be used in creating consistent HTML templates across the application. In the context of this class, we also define important HTML elements such as HEAD and BODY along with meta tags for charset, viewport, description, author, keywords and also link the main stylesheet. This not only leads to more organized code structure but also enhanced maintainability.