Update dependencies and add templating module to build
Updated 'kotlin-css-jvm' to 'kotlin-css' with newer pre-release version and added ':templating' project as a dependency in build.gradle.kts. The change in 'kotlin-css' was necessary to ensure latest features and bug fixes. Additionally, the ':templating' dependency was added to access and utilize the template classes for the project, enabling better code reusability and simpler HTML templates structure.