Remove app.css link from BaseTemplate.kt
Removed the reference to the stylesheet "app.css" in the BaseTemplate.kt file. It was decided that styling should be defined and loaded at a more specific level for increased flexibility. As the project evolved, it became apparent that having a single global stylesheet was Creating constraints and preventing from tailoring styles to individual components efficiently.