Remove 'templating' module from the project
This commit removes the 'templating' module from the project by deleting its .gitignore, build.gradle.kts and CssValue.kt files. This module is discontinued as it was specific to a now obsolete feature. It is also removed from the settings.gradle.kts and build.gradle.kts files to clean up any residual references.