Add website header along with burger menu animation
Introduced a new header section with namespace and burger menu. The burger menu consists of a Lottie animation that toggles between open and close states on click. Additionally, updated the css (AppCss.kt) to style the header and the burger icon. A new string utility function was added to ease handling of strings. For the burger animation, necessary scripts were loaded in SiteTemplate.kt and an external json file was used to store the animation data. Adjustment were also made in Routing.kt and NotFound.kt to complement the new changes.