Refactor codebase to implement burger menu
Refactored the code base to implement a burger menu and enhance the overall design of the website. Multiple CSS files were updated and new ones were added to handle the styles for the drawer, text, and backdrop of the website. Also, created a new Body.appDrawer() function to handle the drawer and its items. Lottie animations were used to handle the opening and closing of the burger menu. AppJs.kt file was added to manage all related app scripts. Adjustment to the rendering of the website's header were made to complement these changes. This enhancement was done to improve the user interface of the website and provide better navigation.